+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 - Improved wizard of new module: Added questions for connection and alternative connection.
 - Added the in-editor validation of `makecomapp.json` project file.
 - Fixed the issue `"options.maxOutputLength" is out of range` occured during ZIP import.
+- Added "Apply for all" into component mapping alignment process.
 
 2.0.0 [2024-07-02]
 --------------------